Re-enable jenkinsfile, add develop/master logic and override
